refactor code to eliminate unnecessary code files
