+++ data/index.ts
@@ -1,7 +1,6 @@
 export * from './useSections';
 export * from './useNavigation';
-export * from './useRecentArticles';
 export * from './useAboutArchwayArticles';
 export * from './useSectionCategories';
-export * from './useSectionPopularArticles';
+export * from './useHighlightedArticles';
 export * from './useAlgoliaSearch';
