+++ pages/community.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
-  const community = "community";
+  import { SortingReplicas } from '@/domain/AlgoliaSearch';
 </script>
 
 <template>
   <div>
     <SectionsHero />
     <SectionsCategories />
-    <SectionsPopularArticles :root-section="community" />
+    <UiPopularArticles :sorting-replica="SortingReplicas.DocsByViewed" title="Most popular articles" css-class="grid grid-cols-1 pt-16 space-y-8 w-full" root-section="community" />
   </div>
 </template>
